refactor chat room and messages
