+++ src/components/ChatDialog/ChatRoom/ChatRoom.scss
@@ -1,0 +1,17 @@
+.no-result {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  height: 400px;
+  justify-content: center;
+
+  img {
+    height: 150px;
+    padding: 0;
+  }
+
+  p {
+    font-size: 15px;
+    line-height: 22.4px;
+  }
+}
