--- src/components/ChatDialog/Conversations/Conversations.scss
@@ -1,117 +1,0 @@
-.conversations-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  .conversations-list {
-    align-items: center;
-    max-height: 383px;
-    overflow-x: hidden;
-    overflow-y: auto;
-    width: 100%;
-    &::-webkit-scrollbar {
-      background-color: #f5f5f500;
-      height: 8px;
-      width: 3px;
-    }
-
-    &::-webkit-scrollbar-thumb {
-      background-color: rgba(0, 0, 0, 0.2);
-      border-radius: 4px;
-      opacity: 0;
-      transition: opacity 0.2s ease-in-out;
-    }
-
-    &:hover::-webkit-scrollbar-thumb {
-      opacity: 1;
-    }
-  }
-  .conversation-grid {
-    align-items: center;
-    column-gap: 16px;
-    cursor: pointer;
-    display: flex;
-    padding: 8px;
-
-    .unread-dot-visible {
-      color: #fa9413;
-    }
-
-    .unread-dot-hidden {
-      opacity: 0;
-    }
-
-    &:hover {
-      background-color: #f29d3456;
-      border-radius: 8px;
-    }
-
-    .thumbnail-right {
-      column-gap: 8px;
-      display: flex;
-      flex-grow: 1;
-      justify-content: space-between;
-
-      .dot-grid {
-        align-items: center;
-        display: flex;
-        margin-right: -8px;
-
-        p {
-          align-self: flex-end;
-          font-size: 11px;
-          margin: 0;
-        }
-      }
-    }
-
-    .thread-details-grid {
-      align-items: flex-start;
-      display: flex;
-      flex-grow: 1;
-      flex-direction: column;
-      row-gap: 4px;
-
-      .unread-message {
-        font-weight: 600;
-      }
-
-      h5 {
-        font-size: 16px;
-        font-weight: 400;
-        line-height: 24px;
-        margin: 0;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        white-space: nowrap;
-      }
-
-      p {
-        font-size: 14px;
-        margin: 0;
-        max-width: 190px;
-        overflow: hidden;
-        text-align: left;
-        text-overflow: ellipsis;
-        white-space: nowrap;
-      }
-    }
-  }
-
-  .no-results {
-    align-items: center;
-    display: flex;
-    flex-direction: column;
-    height: 400px;
-    justify-content: center;
-
-    img {
-      height: 150px;
-      padding: 0;
-    }
-
-    p {
-      font-size: 15px;
-      line-height: 22.4px;
-    }
-  }
-}
